% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ci_upper}
\alias{ci_upper}
\title{Confidence interval (upper 97.5\%)}
\usage{
ci_upper(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector}

\item{na.rm}{A logical value indicating whether NA values should be stripped
before the computation proceeds.}
}
\value{
97.5% quantile of x
}
\description{
Confidence interval (upper 97.5\%)
}
\examples{
x <- rnorm(1000, mean = 0, sd = 1)
ci_upper(x)
}
