% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyboot.R
\name{multi_boot}
\alias{multi_boot}
\title{Non-parametric bootstrap with multiple sample statistics}
\usage{
multi_boot(data, ...)
}
\arguments{
\item{data}{A data structure containg the data to bootstrap.}

\item{...}{Additional arguments passed to particular methods.}
}
\description{
\code{multi_boot} is a generic function for bootstrapping on various data
structures. The function invokes particular methods which depend on the class
of the first argument.
}
\examples{
## List of available methods
methods(multi_boot)
}
