% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyboot.R
\name{tidyboot}
\alias{tidyboot}
\title{Non-parametric bootstrap with multiple sample statistics}
\usage{
tidyboot(data, ...)
}
\arguments{
\item{data}{A data structure containing the data to bootstrap.}

\item{...}{Additional arguments passed to particular methods.}
}
\description{
\code{tidyboot} is a generic function for bootstrapping on various data
structures. The function invokes particular methods which depend on the class
of the first argument.
}
\examples{
## List of available methods
methods(tidyboot)
}
