% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyboot.R
\name{tidyboot_default}
\alias{tidyboot_default}
\title{Nonparemetric Bootstrap and Empirical central tendency for data frames
Designed to make standard use of \code{tidyboot.data.frame} easier}
\usage{
tidyboot_default(data, column, na.rm = NULL, empirical_function = mean,
  statistics_functions = list(ci_lower = ci_lower, ci_upper = ci_upper),
  nboot = 1000)
}
\arguments{
\item{data}{A data frame.}

\item{column}{A string indicating the column of \code{data} to bootstrap}

\item{na.rm}{A logical indicating whether NAs should be dropped before
bootstrapping (defaults to \code{NULL})}

\item{empirical_function}{a string indicating the function to compute and
bootstrap over (defaults to \code{mean})}

\item{statistics_functions}{A vector of strings that are names of functions
to be bootstrapped (defaults to \code{c("ci_lower", "ci_upper")})}

\item{nboot}{The number of bootstrap samples to take (defaults to \code{1000}).}
}
\description{
Computes arbitrary bootstrap statistics on univariate data.
NOTE: Both empirical functions and bootstrapping functions will be computed
 over the grouping variables currently specified for the data frame.
}
\examples{
## Mean and 95\% confidence interval for 1000 samples from two different normal distributions
require(dplyr)
gauss1 <- data.frame(value = rnorm(1000, mean = 0, sd = 1), condition = 1)
gauss2 <- data.frame(value = rnorm(1000, mean = 2, sd = 3), condition = 2)
df <- bind_rows(gauss1, gauss2) \%>\%
 group_by(condition)
 tidyboot_default(data = df, column = value)
}
